添加获取用户信息的接口

+++ app/src/main/java/com/okatu/rgan/blog/controller/BlogController.java
@@ -1,14 +1,20 @@
 package com.okatu.rgan.blog.controller;
 
-import com.okatu.rgan.blog.model.BlogEditParam;
+import com.okatu.rgan.common.exception.ConstraintViolationException;
+import com.okatu.rgan.common.exception.ResourceAccessDeniedException;
+import com.okatu.rgan.blog.model.BlogDTO;
+import com.okatu.rgan.blog.model.param.BlogEditParam;
 import com.okatu.rgan.blog.model.entity.Blog;
 import com.okatu.rgan.blog.model.entity.Tag;
 import com.okatu.rgan.blog.repository.BlogRepository;
-import com.okatu.rgan.blog.exception.EntityNotFoundException;
+import com.okatu.rgan.common.exception.EntityNotFoundException;
 import com.okatu.rgan.blog.repository.TagRepository;
+import com.okatu.rgan.user.model.RganUser;
+import com.okatu.rgan.user.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.client.ResourceAccessException;
 
 import java.util.LinkedHashSet;
 import java.util.List;
@@ -24,6 +30,9 @@ public class BlogController {
     @Autowired
     private TagRepository tagRepository;
 
+    @Autowired
+    private UserRepository userRepository;
+
     @GetMapping
     List<Blog> all(){
         return blogRepository.findAll();
@@ -35,37 +44,49 @@ public class BlogController {
 //    }
 
     @PostMapping
-    Blog add(@RequestBody BlogEditParam blogEditParam){
+    BlogDTO add(@RequestBody BlogEditParam blogEditParam, @AuthenticationPrincipal RganUser user){
         LinkedHashSet<Tag> tags = findTagsByTitles(blogEditParam.getTags());
         Blog blog = new Blog();
         blog.setTitle(blogEditParam.getTitle());
         blog.setContent(blogEditParam.getContent());
         blog.setTags(tags);
-        return blogRepository.save(blog);
+        blog.setUser(userRepository.findByUsername(user.getUsername())
+            .orElseThrow(() -> new ConstraintViolationException("Cannot find current request username in repository, username: " + user.getUsername()))
+        );
+
+        return BlogDTO.convertFrom(blogRepository.save(blog));
     }
 
     @GetMapping("/{id}")
-    Blog one(@PathVariable Long id){
-        return blogRepository.findById(id).
-                orElseThrow(() -> new EntityNotFoundException("blog", id));
+    BlogDTO one(@PathVariable Long id){
+        return blogRepository.findById(id).map(BlogDTO::convertFrom)
+            .orElseThrow(() -> new EntityNotFoundException("blog", id));
     }
 
     @PutMapping("/{id}")
-    Blog edit(@PathVariable Long id, @RequestBody BlogEditParam blogEditParam){
+    BlogDTO edit(@PathVariable Long id, @RequestBody BlogEditParam blogEditParam, @AuthenticationPrincipal RganUser user){
 
-        return blogRepository.findById(id).map(
-            blog -> {
-                blog.setTitle(blogEditParam.getTitle());
-                blog.setContent(blogEditParam.getContent());
-                blog.setTags(findTagsByTitles(blogEditParam.getTags()));
-                return blogRepository.save(blog);
-            }
-        ).orElseThrow(() -> new EntityNotFoundException("blog", id));
+        Blog blog = blogRepository.findById(id).orElseThrow(() -> new EntityNotFoundException("blog", id));
+
+        if(!blog.getUser().getId().equals(user.getId())){
+            throw new ResourceAccessDeniedException("you have no permission to edit this blog");
+        }
+
+        blog.setTitle(blogEditParam.getTitle());
+        blog.setContent(blogEditParam.getContent());
+        blog.setTags(findTagsByTitles(blogEditParam.getTags()));
+
+        return BlogDTO.convertFrom(blogRepository.save(blog));
     }
 
     @DeleteMapping("/{id}")
-    void delete(@PathVariable Long id){
-        blogRepository.deleteById(id);
+    void delete(@PathVariable Long id, @AuthenticationPrincipal RganUser user){
+        blogRepository.findById(id).ifPresent(blog -> {
+            if(!blog.getUser().getId().equals(user.getId())){
+                throw new ResourceAccessException("you have no permission to delete this blog");
+            }
+            blogRepository.deleteById(id);
+        });
     }
 
     private LinkedHashSet<Tag> findTagsByTitles(LinkedHashSet<String> titles){
